backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
django-mailer
django-cors-headers
django-filter
dj-database-url
whitenoise
gunicorn
python-dotenv
requests
